+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Create basic files for `lifespan-development` recipe
+
 ## [v2.6.0] - 2023-11-17
 
 * Add `boxed-feature` note in `neuroscience`